ConioB (buffered) still not working - dead end for using FIFO macros from cbuf.h.
This is last commit before branch which will continue on different principle for the ConioB.
The problem is that the UART driver does not know about the FIFO, it expects normal buffer.
If we store user string to FIFO, it may go from the end back to the beginning and we would need
to call UART.Send twice - once with part of the FIFO from current push position to the end and
2nd time for the remaining string from the start of FIFO. This is hard to do.
Overall the use of FIFO macros seems like dead-end road since the UART driver does not
use the FIFO; the logic of presenting the data to the driver is complicated.